added echo statements to runner script. Lots of file checks to see if things have been done before re-attempting them (have these FAST5 files been basecalled already? e.g. are they already present in the OUTDIR?

+++ scripts/01_basecall-w-gpu.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 # UGE options removed, since this must be performed manually on node 98
 # May eventually be added, if GPUs (Tesla V100s) are available via UGE/qsub
+
+#This function will check to make sure the directory doesn't already exist before trying to create it
+make_directory() {
+    if [ -e $1 ]; then
+        echo "Directory "$1" already exists"
+    else
+        mkdir $1
+        echo "Directory "$1" has been created"
+    fi
+}
+
 set -e
 
 source /etc/profile.d/modules.sh
@@ -28,8 +39,9 @@ hostname
 # Setup tempdir in /tmp
 # Cory recommended this since it will be faster than NFS GWA storage, lower latency as well
 tmpdir=$(mktemp -p /tmp/pjx8/ -d guppy.gpu.XXXXXX)
-trap ' { echo "END - $(date)"; rm -rf $tmpdir } ' EXIT
-mkdir $tmpdir/log
+# rm -rf $tmpdir removed so that it doesn't delete files - TODO add back later
+trap ' { echo "END - $(date)"; } ' EXIT
+make_directory $tmpdir/log
 echo "$0: temp dir is $tmpdir";
 
 # no module load for guppy, since it's installed natively on node 98
@@ -39,7 +51,12 @@ module load qcat/1.0.1
 # should return version 3.0.3
 guppy_basecaller -v
 # basecalling
-guppy_basecaller -i $FAST5DIR -s $tmpdir/fastq --gpu_runners_per_device 48 --num_callers $NSLOTS --flowcell FLO-MIN106 --kit SQK-LSK109 --qscore_filtering 7 --enable_trimming yes --hp_correct yes -r -x auto --chunks_per_runner 96
+# check to see if basecalling has been done by checking OUTDIR for sequencing_summary.txt
+if [[ -e $OUTDIR/fastq/sequencing_summary.txt ]]; then
+  echo "FAST5 files have already been basecalled. Skipping."
+else
+  guppy_basecaller -i $FAST5DIR -s $tmpdir/fastq --gpu_runners_per_device 50 --num_callers $NSLOTS --flowcell FLO-MIN106 --kit SQK-LSK109 --qscore_filtering 7 --enable_trimming yes --hp_correct yes -r -x auto --chunks_per_runner 96
+fi
 
 #### Commented out to try qcat instead
 # Demultiplex.  -r for recursive fastq search.
@@ -50,10 +67,19 @@ qcat --version
 # Demultiplex with qcat, multithreading not supported yet, already really fast on 1 thread
 # Native barcoding kit barcodes 1-24 specified
 # look in both pass and fail folders for reads (poor quality reads to be filterd out later w filtlong)
-cat $tmpdir/fastq/*/*.fastq | qcat --trim -k NBD104/NBD114 -b $tmpdir/demux --detect-middle 
+if [[ -e $OUTDIR/demux/none.fastq ]]; then
+  echo "FASTQ files have already been demultiplexed. Skipping."
+else
+  echo "Demultiplexing with qcat now..."
+  cat $tmpdir/fastq/*/*.fastq | qcat --trim -k NBD104/NBD114 -b $tmpdir/demux --detect-middle 
+fi
 
 # retain the sequencing summary
-ln -v $tmpdir/fastq/sequencing_summary.txt $tmpdir/demux
+if [[ -e $tmpdir/demux/sequencing_summary.txt ]]; then
+  echo "sequencing_summary.txt has already been hardlinked into demux/"
+else
+  ln -v $tmpdir/fastq/sequencing_summary.txt $tmpdir/demux
+fi
 
 #### Commented out because qcat doesn't produce separate directories per barcode
 # Make a relative path symlink to the sequencing summary
@@ -62,13 +88,20 @@ ln -v $tmpdir/fastq/sequencing_summary.txt $tmpdir/demux
 #  ln -sv ../sequencing_summary.txt $barcodeDir/;
 #done
 
-cp -rv $tmpdir/demux $OUTDIR
+if [[ -e $OUTDIR/demux/none.fastq ]]; then
+  echo "Demuxed fastqs have been transferred from tmpdir to OUTDIR. Skipping."
+else
+  cp -rv $tmpdir/demux $OUTDIR
+fi
 
 # create subdirectories for each barcode, move fastq file there (does not do so for none.fastq)
-for fastq in $OUTDIR/demux/barcode*.fastq; do
-  mkdir $OUTDIR/demux/$(echo $fastq | cut -d '/' -f 3 | cut -d '.' -f 1)
-  mv $fastq $(echo $fastq | cut -d '.' -f 1)
-done
-
+if [[ -e ${OUTDIR}demux/barcode01/ ]]; then
+  echo "fastqs assigned a barcode have already been moved into subdirs. Skipping"
+else
+  for fastq in ${OUTDIR}demux/barcode*.fastq; do
+    mkdir ${OUTDIR}demux/$(echo ${fastq} | cut -d '/' -f 3 | cut -d '.' -f 1)
+    mv $fastq $(echo $fastq | cut -d '.' -f 1)
+  done
+fi
 # making OUTDIR available to runner script for copying logfile into $tmpdir/log
-export $OUTDIR
+export OUTDIR
